Resort store response set on internal label dedup
When deduplicating on labels which are stored internally in TSDB,
the store response set needs to be resorted after replica labels are removed.

In order to detect when deduplication by internal labels happens, this PR adds a
bloom filter with all label names to the Info response. When a replica label is present
in this bloom filter for an individual store, the proxy heap would resort a response set from
that store before merging in the result with the rest of the set.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>