Remove retriable query implementation
We suspect the retriable query could be causing nil pointer exceptions.

Since we do not know how often we trigger the retry in the first place,
it is better to remove it and re-add if we see a necessity to do it.

The implementation is not part of the upstream version so this would
only bring us closer to removing the fork.