Cache query results
We have a few dashboards which use the same expressions multiple times inside a single query,
which can overwhelm regional queriers, leading to timeouts.

This commit adds caching capabilities to the remote engine implementation so that
query results are cached within the context of a single query invocation.
Ideally this optimization should eventually end up in the engine directly.