Reduce copies
Calling params.SetWr seems to make a copy of the field instead of setting
a reference.

This commit optimizes the call to set an empty message, and then build the payload
into it. Since Cap'n Proto is asynchronous, the arena release has to happen after we've
received a response from the server.