Optimize marshaling
Storing symbols as a list of Text items seems to be expensive
because each text item is treated as a separate object.

We can optimize the layout by storing a byte array of all symbols
alongside a list of offsets pointing to the first character for each symbol.
Reading the table becomes a matter of doing a byte range lookup using
the current offset as start and next offset as end.