Decouple receiver writers
Currently the receiver writer takes in an interface to a write request.
This makes it hard to make optimizations based on the request protocol (proto, capnproto)
since everything has to conform to the same interface.

This commit decouples the two writers so that we have one writer for each protocol.