Avoid using protobuf structs
The capnproto writer is still coupled to protobuf structs.
This makes it hard to use stringlabels and also has incompatibilities
with the upstream version.

This commit introduces independent types which are tightly coupled
to the prometheus model. In addition to decoupling from proto, we can
now unmarshal directly into prometheus data structs and avoid unneeded allocs.